split getFieldsNotInSubschema into separate file
like mergeFields
